Restructured the whole code base
