+++ Zoop/Placement/ZoopBuildExecutor.cs
@@ -1,8 +1,9 @@
 using System.Reflection;
 using Assets.Scripts.Inventory;
 using Assets.Scripts.Objects;
+using ZoopMod.Zoop.Core;
 
-namespace ZoopMod.Zoop;
+namespace ZoopMod.Zoop.Placement;
 
 internal static class ZoopBuildExecutor
 {
