+++ Zoop/Placement/ZoopSmallGridCoordinator.cs
@@ -6,8 +6,11 @@ using Assets.Scripts.Objects.Pipes;
 using Assets.Scripts.Util;
 using Cysharp.Threading.Tasks;
 using UnityEngine;
+using ZoopMod.Zoop.Core;
+using ZoopMod.Zoop.Planning;
+using ZoopMod.Zoop.Preview;
 
-namespace ZoopMod.Zoop;
+namespace ZoopMod.Zoop.Placement;
 
 /// <summary>
 /// Owns the small-grid zoop preview flow from path planning through preview instantiation and placement.
